refactor(harness-synth): split synthesize.ts into focused modules
The synthesize.ts file grew to 1208 lines during the adversarial review
loop as defensive layers (input validation, JSON-plain checks, guard
attempts, verifier coercion) accumulated. This pushed it past the 800-
line hard limit and the complexity ratchet to 743 (CI cap: 742).

Extract helpers into focused siblings:

- json-plain.ts (219 lines) — snapshotJsonPlain, ensureJsonPlain, jsonEqual,
  deepFreeze, MAX_JSON_DEPTH. Single home for the JSON-plainness trust
  boundary used by every input-validation surface.
- guarded-attempt.ts (203 lines) — GuardedResult type, linkSignal,
  guardAttempt, safeSanitize, redactReason. Per-attempt timeout/abort
  primitive plus the sanitization helpers that protect prompt/public
  surfaces.
- verify-coerce.ts (183 lines) — CoercedVerifyResult, freezeDescriptor,
  coerceVerifyResult, coerceVerificationSummary. Verifier-output
  validation; coerceVerifyResult split into a coerceOkTrue helper to
  satisfy the 50-line function limit.
- input-validation.ts (163 lines) — validateAndSnapshotInput plus its
  helpers. Trust-boundary entry that replaces the inline pre-loop checks.
- safe-callbacks.ts (95 lines) — safeGenerate, safeVerify; map guarded
  failures to structured SynthesisFailureKind values.

synthesize.ts is now 391 lines, well under the soft limit. The main
function delegates to validateConfig, validateAndSnapshotInput, and
runSynthesisLoop; the loop itself delegates per-attempt work to runAttempt
and runVerifyStage, which return a small AttemptOutcome discriminator the
loop reads to update lastReason/lastKind/priorCode/priorReason.

No behavior changes — same 107 tests pass; same 15/15 E2E matrix passes.
Complexity ratchet: 743 -> 742 (within CI budget).
